requirements.txt
alive-progress
requests
python-dotenv
